requirements.txt
discord.py==1.7.3
PyYAML==6.0
torch==1.10.2
transformers==4.18.0
